[#4] chore: file 분리하고 vertical stackview가 되도록 함
