+++ iOS/BaseballGame/BaseballGame/Views/GameInfoStackView.swift
@@ -1,0 +1,27 @@
+//
+//  GameInfoStackView.swift
+//  BaseballGame
+//
+//  Created by kimdo2297 on 2020/05/04.
+//  Copyright © 2020 Jason. All rights reserved.
+//
+
+import UIKit
+
+final class GameInfoStackView: UIStackView {
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        translatesAutoresizingMaskIntoConstraints = false
+        configure()
+    }
+    
+    required init(coder: NSCoder) {
+        super.init(coder: coder)
+        configure()
+    }
+    
+    private func configure() {
+        axis = .vertical
+        distribution = .fillEqually
+    }
+}
